[MINOR] fix check style error
