+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestCOWDataSource.scala
@@ -790,17 +790,4 @@ class TestCOWDataSource extends HoodieClientTestBase {
     val resultSchema = new StructType(recordsReadDF.schema.filter(p=> !p.name.startsWith("_hoodie")).toArray)
     assertEquals(resultSchema, schema1)
   }
-
-  @ParameterizedTest
-  @ValueSource(booleans = Array(true, false))
-  def testWithEmptyInput(allowEmptyCommit: Boolean): Unit = {
-    val inputDF1 = spark.read.json(spark.sparkContext.parallelize(Seq.empty[String], 1))
-    inputDF1.write.format("org.apache.hudi")
-      .options(commonOpts)
-      .option(DataSourceWriteOptions.OPERATION_OPT_KEY.key(), DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL)
-      .option(HoodieWriteConfig.ALLOW_EMPTY_COMMIT.key(), allowEmptyCommit.toString)
-      .mode(SaveMode.Overwrite)
-      .save(basePath)
-    assertEquals(allowEmptyCommit, HoodieDataSourceHelpers.hasNewCommits(fs, basePath, "000"))
-  }
 }
